chore(neon): delete the reconciler, the parity sweep and the mirror-lag watchdog
All three exist only to hold D1 and Neon against each other, and Neon is the
sole store for every table now. The reconciler has been a no-op since the lanes
finished; the parity sweep counts a frozen store and reports a permanent,
growing deficit; the mirror-lag watchdog measures the lag of a copy nothing
writes. ~5,800 lines, three cron triggers and one flag.

NEON_BACKFILL_LANES goes with them rather than being left declared-and-empty.
An empty flag reads as "no tables need this" instead of "this question no
longer exists", and two prune gates had already inverted under exactly that
ambiguity (#10152, #10164) -- leaving it would be leaving the trap.

Wiring unpicked:

  assertIdentifier moves to neon-write.ts, where the surviving caller
  (neon-prune) already imports from.

  chain-detail-prune's Neon gate was `lanes.has(table) || neonOwnsTable(...)`.
  The first arm was the reconciler's, so the gate is ownership alone.

  neon-sql-portability derived its boolean deny-list from
  NEON_BACKFILL_PLANS[*].booleans. The authority moves to the surviving write
  plans, and the check becomes "every declared column is guarded" rather than
  set equality -- the deny-list also covers columns whose only writer is a
  storeBoolean call, and dropping those because no plan names them is how one
  goes unguarded.

  observations-flip-readiness listed all three under STORE_AWARE.

Full suite: 763 files, 18,145 passing.

Closes #10166
